Fix: some sokoban levels could easily be flagged as 'never solved'.
There are a couple sokoban levels that were added fairly recently, that
the player did not have to fill every single pit to solve the level. The
problem is, in maybe_finish_sokoban() it checks for all pits on the
level. If any remain unfilled, that level is not considered solved.
Meaning luck penalties could still apply for various actions, moving
diagonally past a boulder was still not possible, etc. So to fix, tweak
the offending maps so this doesn't happen, and without changing the
difficulty of the level.