Fix: refusing to hand over the quest artifact to your quest leader was still flagging the quest as complete.
A loophole that was being taken advantage of - player has abused their
alignment, player brings the quest artifact back to their quest leader,
and it's determined that the quest leader wants the quest artifact
returned. At this point is where the quest was flagged as complete. The
player could then evade the quest leader and escape, and finish the
game. I've known about this loophole for awhile, but it wasn't a
priority to address until now. As of this commit, if the quest leader
demands that the quest artifact be returned and the player does not
comply, the quest is no longer flagged as complete. They'll have to
defeat their quest leader to complete the quest and be able to perform
the invocation.