Include race in statues of previous players
Instead of creating 'a statue of a wizard named K2' make it 'a statue of
a gnomish wizard named K2', using the race information from the top ten
file.

Unfortunately, the current top ten format being used by evilhack is the
deprecated version that predates separation between class and race
selection, so it doesn't include race.  As a result, this patch will not
actually produce an immediately noticeable difference -- but once the
version number increases, it will switch over to start using the modern
record format, and any statues of players from version 0.7.2+ will
include their race.

Normally the changeover to the new top ten record file format would
happen once we reach EvilHack 3.3.0, because that's where it changed in
vanilla NetHack; I updated those criteria in the patch so that EvilHack
will start using the modern format as of 0.7.2.