Adjust minimum distance a monster will cast fire/ice bolt or acid blast.
Typically a spellcasting monster will not nuke the player with fire/ice
bolt or acid blast if they are too close, as they could wind up in the
blast radius. Discovered that monsters would still cast these spells if
they were diagonally adjacent to the player:

  x.x
  .@.
  x.x

If a monster was standing on a tile represented by a . they wouldn't
cast, but if standing on x, they would. So I've tweaked the minimum
distance so that the x spaces are considered.

Another fun tweak - spellcasting monsters will follow the above rule
most of the time, but I've now introduced a small chance that they'll
YOLO it and cast while standing right next to the player.