Fix: no option to repair an attached rusted/corroded ball & chain.
Was noticed when a player was trying to repair their starting ball &
chain (still attached), and had no intent of trying to be free of it
yet. The reminder that you'd need a hammer to break free of the chain
will still occur (sometimes now instead of every time), but now the
remaining 'dipped what into a forge' routine will be called.