Fix: black colored tame monsters not showing in the HTML dumplog.
This was noticed during the TNNT 2021 tournament. In the HTML dumplog,
any tame monsters that were colored black would not appear, as their
color was the same shade of black as the background.

Also found and fixed an omission in Makefile.top, the NHdump.css file
wasn't being copied over to its install directory via 'make install'.