Make destroy armor vs mon match effect vs hero
Destroy armor cast against a monster would instantly destroy a piece of
armor, and could be resisted by high level monsters; destroy armor
against the hero would erode a piece of armor to a certain extent, and
then destroy it outright after that point.  Make destroy armor work the
same way vs a monster as it does vs the hero.