Fix: exploding acid spheres and damage.
A couple things... first, it was noticed that if the player had an
object protected by grease, that grease would dissolve if an acid sphere
exploded, regardless of its proximity to the player. Solution: move acid
damage effects to their proper location in explode() and remove it from
explmu(). This way (the second bit), any type of acid explosion will
behave as it should, and not just from an acid sphere monster explosion.