Fix: edit stock sys/unix/hints/linux file for installation use.
Most people won't think to look for/use the linux-debug hints file, so
lets edit the stock linux hints file so it actually works. After
compiling with this, the game will throw a warning about LIVELOGFILE not
defined, which can be ignored - easy fix for that if you're OCD like I
am, is to uncomment the CFLAGS values for DDGAMELAUNCH and
DLIVELOG_ENABLE (which are not necessary for local play at all).