Apply a wielded polearm with fire-command.
From NetHack 3.7 git commit fa41d5f (modified):

'Wield a polearm and use 'f'ire to automatically hit with it,
if there's a single valid target. This only applies if quiver is empty
and autoquiver is off.'

Autoswapping to a polearm at the ready was not included in this commit,
I don't feel this would always be desireable. If the player has a
polearm at the ready and wants to swap to it, they can do so manually.

Using 'f'ire to autohit a valid target, the player would need to be the
same distance away from its intended target as they would when
'a'pplying their polearm.