Adjust alignment record penalty for Convict's killing 'always peaceful' monsters that became hostile just from seeing the Convict.
Convict players cause several 'always peaceful' monster types to
automatically become hostile on sight. Many times they have to kill
these monsters in self defense, but were still taking the same severe
alignment penalty for killing them. So we tweak it in this commit -
there's still an alignment penalty (our Convict should learn to evade,
not engage), but it's nowhere near as bad. On the order of 2 points
instead of 15.