Indestructable objects have same chance to resist cancellation as intelligent artifacts.
Those being the invocation items, the Amulet of Yendor, and the Idol of
Moloch (if playing Infidel role).

Also some tweaks and improvements - don't count intelligent artifacts or
indestructable objects when checking for blessed objects. If the random
objects to be cancelled count is greater than the number of objects
selected, clamp the cancellation count to be no greater than the
selection count (cuts down on unnecessary spam).