Block 'artificial' player monster creation
Restrict creation of player monsters to 'natural' monster generation
during level creation and over the course of the game, preventing them
from being created by the create monster spell/scroll/wand, a bag of
tricks, etc.  Player monsters are supposed to be other adventurers
exploring the dungeon, and a fully kitted-out, high-level priestess
popping out of a magical bag (or being summoned by the player or a
spellcasting monster, etc) seems inconsistent with that notion.