Fix: shambling horror generating with different stats on bones file load.
The issue: player starts a game, a shambling horror for their particular
games has a set type and number of stats. Player then finds a bones
level - any future shambling horrors generated now had the stats for the
game that the bones file originated from. This was not the intent. This
commit fixes this, so once a game is started, whatever the stats are
going to be for a shambling horror for that game are locked in and will
not change. Thanks qt for making this look easy to fix.

Co-authored-by: entrez <me@entrez.cc>