Fix: invalid material 7 (otyp 427).
Would encounter this whilst fuzzing. otyp 427 is a spellbook of blank
paper, material type 7 is leather. There are two types of spellbooks
that are considered being made out of leather - vellum and parchment.
There was already a fix in place if you had a spellbook made of leather
and it was blanked from a water damage, but the same treatment was
missing should those same types of spellbooks become cancelled, or read
too many times and fade away.

There may be another case I'm missing, but this at least addresses two
of them.