Fix: unnecessary impossible() in restmonsteeds().
Came across this whilst fuzzing - in restmonsteeds(), if a monster
steed (mtmp) was not found, then free it up (free_erid()). But there was
an impossible() line right before that, which would cause an error, and
never reach the function free_erid(). That line doesn't even need to be
there.

Also a bit of code formatting cleanup.