Fix: impossible 'placing <monster> over <monster> at <location>'.
I've seen this 'program in disorder' message in the paniclog quite a few
times across every version of EvilHack. It would almost always occur
when the player was engulfed by some monster, and the game would try to
place the engulfing monster where the player was before being engulfed.
This commit adds a check to make sure the position it wants to move the
engulfing monster to is ok for it to be in. After adding this check and
fuzzing the game repeatedly (up to more than 100 million turns), I
haven't seen this error reappear.

This bug also exists in vanilla NetHack, and I've seen it expressed in
the paniclogs for version 3.7, but it doesn't happen nearly as often, as
vanilla NetHack has much fewer engulfing monsters than EvilHack.