Fix: return DOORMAX to proper value, level creation fixes.
Where do I begin with this one? It was discovered that DOORMAX was
causing a buffer overflow, as it's value was raised to 200 (highest it
can be is 127, initial value was 120). The reason that was increased way
back was because when testing various random vault rooms (themed rooms
as they're now called in NetHack 3.7), the number of doors created was
exceeded. Didn't really investigate why that was happening at the time.
While investigating DOORMAX, qt discovered why so many doors were being
created in the first place (which is addressed in sp_lev.c), literally
multiple doors on the same tile, which could adversely affect corridor
creation and how rooms can connect to each other. This is turn led to a
clean up of a couple of random vault rooms which were generating rooms
inside a larger room (should be subroom inside a normal room). Bleh.

Co-authored-by: gebulmer <gebulmer@googlemail.com>
Co-authored-by: entrez <me@entrez.cc>