Infidels cannot regenerate hit points without the Amulet of Yendor in their possesion (pre-imbued Idol of Moloch).
A while back, a deterrent was added to prevent Infidels from dropping
the Amulet of Yendor and bypass its effects - blood magic. Without the
Amulet in their possession, if an Infidel cast spells, it would consume
hit points instead of spell energy (healing spells would cost
substantially more). It was recently brought to light that this was not
a decent deterrent, as the player could simply wait a few turns and let
natural hit point regeneration do its thing, and if the player also had
a ring of regeneration on? Pretty much counteracts 'blood magic'.
completely. The suggestion was made to have blood magic cost both spell
energy as well as hit points, but I think a better deterrent is making
it so the Infidel can't regenerate hit points at all if they don't have
the Amulet in their possession.

Just as with blood magic, this penalty goes away once the Infidel has
sacrificed the Amulet of Yendor and imbues the Idol of Moloch.