Fix: monsters benefit from resistances imparted by worn/wielded gear.
A big part of this commit comes from NetHack 3.7 git commit 439b6b7, as
well as a subsequent commit right after it to address a small bug that
crept in. The problem was that monsters would not benefit from the
secondary extrinsics imparted by worn dragon scales/scale mail. Players
polymorphed into a dragon were also not gaining the benefit of those
secondary extrinsics. Some of this was already addressed here, but not
all the way, there were plenty of gaps. Example: a monster wielding Fire
Brand did not enjoy the fire resistance it would impart to a player
wielding it. In the NetHack commit, there were several areas where
defended() should have been added, but wasn't. In this commit, I've
tried to be as thorough as possible in covering every instance where
that function should be. A second or third pair of eyes on this wouldn't
be a bad idea though. I may have missed a spot or two, or maybe added
the function where it isn't really needed.

While incorporating this, I noticed a couple minor bugs with object
properties, and fixed one of them here. A monster wearing a piece of
armor with the drain resistance property wouldn't work for them - now it
does.

Another bit that made its way into this commit - chromatic dragon
scales/scale mail now emits light just like gold dragon scales/scale
mail do. And because of this, Tiamat herself also emits light.