Fix: preserve the Oracle and Charon's starting stats in bones.
Did this for shambling horrors initially, same needs to be done for the
Oracle and Charon should the player encounter them in a bones level. In
testing, the Oracle, if she wandered around, is not at her starting
place. But she at least stays stationary and has her original stats and
hit points, and is peaceful.