Wand of striking/force bolt spell has a chance to knock back its target.
This is a bit of a 'beta' commit, as I may change the knock back
criteria in the future. The way it currently works is this - if hit with
a wand of striking, or the spell 'force bolt' (player or monster), and
enough damage was generated (greater than 16 from a 2d12 roll), then the
target can be knocked back one tile space. In the future, I may make
some changes, such as force bolt spell would require the caster to be
skilled or better in attack spells, or a blessed wand of striking may
have a lower damage requirement, and/or knock the target back an extra
space.

Monster vs player and monster vs monster works fine. There's one current
bug with player vs monster - player hits monster with wand/spell,
monster gets knocked back one tile, but it acts like the monster is
knocked back before the force bolt travels to the next tile, so they get
hit by the same wand/spell again, get knocked back again, rinse/repeat.
I can prevent this from happening by setting a condition in bhit() so
the range of the wand/spell stops after the knock back, but then no
damage is taken. Will need to figure this one out, but for now I need to
commit what I have so I don't lose track of things.

Thanks qt for the help with the mon vs player bit.