Fix: wand of striking/force bolt spell knocking back its target.
This commit addresses the problem with the target being hit repeatedly
by the same wand/spell if hurtling away. There's a check now to see if
the target has already been hurtled, and if it has, the attack stops.

qt asked me not to mention his name on this commit because he feels his
solution is too much of a kludge. So I won't.