Fix: stack breakage issues.
From the mind of qt:

'In order to make it so that a stack of glass weapons would break one at
a time, breakobj was changed in 89c9bb0 (xnh) to break only a single
item per stack instead of the entire stack at once.  This caused various
problems with other scenarios where stacks might be given to breakobj:
breaking stacks of potions with striking (caused an impossible and
contradictory messages), dropping stacks of potions while levitating
(caused a memory leak as the remaining stack didn't get placed on the
ground, but didn't get freed either), etc.

Instead of breaking stacks item-by-item in breakobj, since the majority
of functions which call it expect it to break an entire stack at once,
restore breakobj to its vanilla behavior and change break_glass_obj to
split the stack if necessary, passing to breakobj only the item which
should be destroyed.  This should preserve the desired 'break one glass
weapon at a time' behavior while fixing the various problems caused by
the breakobj modification.

Co-authored-by: entrez <me@entrez.cc>