Handle air/liquid when choosing mimic disguise
Make it so mimics clinging to the ceiling don't disguise themselves (not
because they can't, but because there isn't really an appropriate
disguise that hangs from the ceiling available in NetHack -- there's no
'archway' furniture or anything like that).  Also make sure air spaces
are permitted as valid positions for flying/floating/clinging monsters.