Fix: partial revert of commit 654b1ce.
Picked this up while fuzzing, wound up with the 'alphabet soup'
impossible. A shambling horror cast a bolt type spell that fell out of
range of what was possible, because of the aforementioned commit.

The fix - the description can still be adjusted so it displays
correctly, but actually adjusting the attack type can lead to issues.
That's the bit that is reverted. So you'll get 'the barbed devil zaps
you with a bolt for fire!' but the hit feedback will show as 'fireball'.
Not quite sure the best way to resolve that last part, so this will do
for now. Not causing an impossible is better than a slightly off
message.