Fix: prevent dmonsfree with a couple mhurtle scenarios.
This could occur when a monster caused another monster to hurtle away.
If the hurtled monster died from the inital cause, and was hurtled into
something else (trap) can tried to kill it again, dmonsfree impossible.

I took bits out of mhurtle_to_doom() to address this. Can't use
mhurtle_to_doom() here, as it assumes the player initiated the monster
hurtle, not an attacking monster.