Fix: impossible 'Webbing over trap type'.
Another one found while fuzzing. This bug has existed for nearly three
years, but the odds of it occuring would be very low. When adding new
traps (spear trap, magic beam trap), I neglected to take those into
consideration should the player polymorph into a type of spider and spin
a web over it. With this fix, the player now has a way to effectively
nullify a spear trap.