The Hand of Vecna.
Previously, upon destroying Vecna, he would maybe drop the Eye of Vecna,
or sometimes he wouldn't. Now when he's destroyed, he'll either drop the
Eye or the Hand of Vecna (50/50 chance). One or the other, but never
both (or neither). This artifact is a special case - in other variants
that have the Hand of Vecna, it's just another food object, much like
the Eye. For EvilHack, it's much more. The Hand of Vecna is technically
a pair of flesh gloves that the player can 'wear'. In inventory it won't
show as a pair of gloves, but as a single mummified hand. When putting
it on, the player receives some rather scary feedback about how their
left hand withers away and dies, and the Hand replaces their original
hand, fusing with their arm. It even shows that way when viewing it in
inventory. Because of this, once the Hand of Vecna is 'worn', it can
never be removed. And because it's technically gloves, the player can't
wear another pair of gloves over it, which is in keeping with how it was
handled via many ad&d sessions, as the Hand of Vecna would not allow
itself to be covered or confined by such simple things.

The breakdown of its abilities and behavior:
* The Hand of Vecna appears as a 'mummified hand', and takes up the
glove slot for worn armor
* Once worn, it cannot be removed, ever. Even if the player polymorphs
into a monster that would normally cause it to drop their gloves
* Because of this, it is immune to erosion, disintegration, destroy
armor monster spell, and destroy armor scroll. It also cannot be stolen
when worn
* It can be enchanted up to +7, but because it can't be destroyed, it
will never evaporate if attempting to enchant it higher than +7 (nothing
happens)
* Special abilities - hungerless regeneration, half physical damage,
disease resistance. It can be invoked for its death magic, killing or
hurting any monster within line of sight that isn't resistant to death
magic. Just as with the Eye of Vecna, invoking its death magic has
consequences
* The Hand of Vecna also confers giant strength like gauntlets of
power, and will cause an additional 1d5 + 7 (8-12) hit points of cold
damage via melee attack to monsters affected by cold, regardless of
whether a weapon is wielded or are fighting bare-handed
* The Hand must be worn for any of the above-listed special abilities to
work
* Worn rings can be removed without issue without needing to remove the
Hand of Vecna, even if it's cursed - it's not treated as a pair of
gloves, even though it techincally is
* Because the Hand is technically dead, players can pick up objects such
as a cockatrice corpse without penalty (another case of 'it's not a pair
of gloves, but technically it is')
* Monsters are scared of the power of the Hand of Vecna, and will not
attempt to 'merge' with it

I may have missed some cases where the player could actually remove the
Hand once worn. If so, should be fun hunting those cases down. This is
another one of those artifacts that was a lot of fun to make, and should
give the player that much more incentive to try and defeat Vecna.

There's a couple small bugs I've noticed and corrected while working on
this, the most important one being that hungerless regeneration was not
working when wielding artifacts like Trollsbane or the Staff of
Aesculapius. Not entirely sure that innate giant race hungerless
regeneration was working either. They do now though.

Enjoy :)