Tortle as a playable race.
Another new race for EvilHack, added at the request of amateurhour.
Tortles are humanoid tortoise-like beings. They are large, and have a
hard shell on their back which they can retreat into when threatened.
They have several armor restrictions due to their size and body shape.
Tortles are typically lawful in alignment, but are sometimes neutral as
well. To get an idea of what a tortle looks like, along with several
other characteristics, check out their Forgotten Realms entry online.

The specifications:

* Tortles can play as either an Archeologist, Barbarian, Healer, Monk,
Priest, Tourist, or a Wizard
* They can max their strength to 19 and wisdom to 20. Dexterity is
capped at 10, charisma is capped at 14
* They can be either gender, alignment choices are lawful or neutral
* Tortles start with the ability to swim, and have magical breathing
(can hold their breath for a ridiculously long time). They gain warning
at experience level 5, and hungerless regeneration at experience level
12
* Tortles start with a naked AC of 0 due to their hard shell, but are
very limited in what kind of armor they can wear
* Tortles cannot wear body armor, cloaks, t-shirts or boots (body
size/shape). They can wear gloves and helms, but only if their material
is made of something that isn't hard or rigid. Any type of shield is
fine, as are rings and amulets
* Tortles are the slowest of the player races, with a speed of 8
* A new type of helm was created - the toque (think Jayne from the show
'Firefly'). Like the fedora, it is made of cloth, does not grant any
extra AC, does not spawn randomly, and will only appear as a tortles
starting gear depending on what role they choose. It can also be wished
for. One special quality it has - it can protect the wearer from sonic
attacks (those goofy earflaps have a function after all)

There's more to do with the Tortle, this is just the initial commit to
get it in. What still needs to be done: giving the tortle the ability to
hide in its own shell, and figure out and implement the pros and cons to
doing that - player monster tortles - stand alone tortles as a monster
the player could encounter. And of course other little tweaks and
changes that haven't been thought of yet.

This should be a fun but challenging race to play once it's complete.