Tortles as player monsters.
Player monsters can now spawn as a tortle given the correct role, and
will have the appropriate armor given what they are.

While doing this, I noticed that a monsters natural AC wasn't carrying
over to player monsters. Not too big of a deal, as most player monster
races have a base AC of 10. Kind of a big deal for giants (base AC of 6)
and an actual big deal for for tortles (base AC of 0). So this has been
addressed as well (thanks qt for pointing me in the right direction).