Tortle ability to hide in its shell.
This commit allows the player as a tortle to hide in its shell via
the #monster command. Once in it's shell, it can stay inside for a set
period of time before it's forced out, or the player can exit early by
issuing the #monster command again.

Once in their shell, the player gains an extra 40 points of AC, but are
immobile and blind, and will be unable to do a several number of things
until they emerge. The benefit of hiding in ones shell, is to be able to
'hunker down' in an attempt to make it through a tough situation that
they otherwise may not be able to survive. Or it could be used as a
simple 'safe space' to rest. It'll be interesting to see how this
ability is used.

Currently, while the player tortle is hiding in their shell, they cannot
do the following: see, move, fight (normally or force attack), cast
spells, use a polearm, engrave, use various tools (towel, stethoscope),
throw objects, #loot things, sacrifice a corpse at an altar, use a
saddle or barding, apply a carried container, or mount a steed. There's
a whole list of other capabilities that will need to be covered as well,
such as: zapping a wand, using other tools (unlocking tool, whistle,
leash, unicorn horn, musical instruments), applying an axe or
pick-axe... and probably many other things that haven't been considered
yet. But this is a good starting point. Being polymorphed while hiding
in your shell needs to be addressed as well. A wait period should also
be imposed between times hiding in ones shell, much like #invoking an
artifact to use its special power, otherwise this could be abused.

More to follow. Thanks qt for getting this started and providing the
core code.

Co-authored-by: entrez <me@entrez.cc>