Implement a cooldown timer after a tortle hides in its shell.
Added a cooldown timeout for tortles hiding in their shell. After they
emerge from it, the player will have to wait several turns before they
can hide in it again. The cooldown period is the same as a prayer
timeout (rnz(350) - see https://nethackwiki.com/wiki/Rnz for a detailed
explanation of this function).

Also added half physical damage whilst hiding in ones shell, and bumped
up the time the player can stay inside before being forced out to 200
turns.

To recap - when a player tortle hides in their shell, they get a 40 AC
boost, and take half physical damage for the duration they are in their
shell. Up to 200 turns max (can emerge earlier if they so choose), and
have to wait rnz(350) turns before they can hide in it again.

Fairly powerful ability in my opinion. This may require further
adjusting - on the surface, this feels too powerful. But with the
inclusion of a cooldown timeout, and the fact that tortles can barely
wear anything armor-wise and can hardly do anything but rest while in
their shells, this may be dialed in already. We'll see. I'm eager to see
how players utilize it.

Thanks qt for implementing the cooldown timeout. Made this look easy.

Co-authored-by: entrez <me@entrez.cc>