Fix: warning: array subscript [0, 1] is outside array bounds of ‘const char * const[2]’ [-Warray-bounds].
This is a vanilla bug, but I'm tired of seeing it every time I compile.
The array 'vampmsg' has only two messages in it for hostile vampires to
use when #chat(ing) them. The way the code was written - if (vampmsg ==
0) do this, else if (vampmsg == 1) do this, else vampmsg[vampindex].
We've already covered the only two states it can be in, the last bit is
implying a third that doesn't exist and is outside the bounds of the
array.