Fix: add message for failed fountain monster detection.
From the inner workings of qt's brain:

'When the fountain quaffing monster detection effect was triggered on a
level without any monsters, no message would be printed.  I think this
was the only scenario where drinking from a fountain wouldn't print
anything, so it stood out as unusual.

Print a messsage in the case monster detection fails, to make it
consistent with other fountain effects and ensure it's clear the hero
did still drink from the fountain on that turn.'

Co-authored-by: entrez <me@entrez.cc>