New branch: Goblin Town.
This new optional side branch introduces Goblin Town, which is ruled
over by the Goblin King. It's designed as an early quest for low-level
characters - a bit challenging, but nothing too crazy or over the top.
The Goblin King is the boss monster at the end of this quest, and he
controls access to Mine Town and Mines' End in the Gnomish Mines.

The particulars:

* Goblin Town is accessible via a magic portal that can be found on
dungeon levels two or three
* Goblin Town is a single level, filled with goblins and hobgoblins
* The Goblin King can be found on his throne at the end of the branch.
His treasure chest contains a few random gems, a random ring, 100 gold
pieces, and there's a 10% chance of a magic marker being present as well
* The Goblin King controls access to Mine Town and Mines' End - if the
player tries to access either place via the stairs or falling/digging
down, they will be blocked from entering. The only way to open up
physical access to both places is to defeat the Goblin King
* The denizens of Goblin Town are always hostile. They have captured a
few elves and keep them as prisoners, who would probably be grateful to
the player if freed
* Just like any other town, Goblin Town has a shop, but there's only a
33% chance that it's actually occupied. The shopkeeper in Goblin Town
will always be an orc

Thought this might be a nice detour for the player during the very early
stages of the game. This branch is purely optional, but a big part of
the draw will be to open up physical access to Mine Town/Mines' End, and
to also have a chance at an early shop or acquire a magic marker. The
Goblin King is based off of the commented out template already found in
monst.c, with several modifications. He only has 8 hit dice and not a
lot of AC - still a concern for a very low-level character, but could be
easily dispatched with a bit of planning.