Player can create various types of armor and weapons using a forge.
Players can now use a forge to create all sorts of different types of
armor and weapons. The bulk of this code comes from SpliceHack, but
I've changed a fair amount of it for EvilHack. The details:

* Two objects can be placed inside a forge to create a new singular
object using the command #forge
* Various recipes exist for almost all types of metallic armor and
weapons
* A hammer (any type) must be wielded to effectively forge an object
* Only metallic objects can be forged
* When selecting the base and secondary objects, the BUC status and
material of the secondary object determines the same for the final
product
* If the base and secondary objects are enchanted, the greater of the
two enchantments carries over to the final product
* Currently, only non-magical weapons and armor can be forged (this may
change in the future)

In SpliceHack, players can forge an assortment of magical objects, and
even artifacts. I felt this was too overpowered, especially since there
isn't an additional cost for doing such a thing. Maybe in the future,
allow magical objects to be forged, but require an additional component
for that to happen.