Fix: priest tortle couldn't wear the Mitre of Holiness.
Not sure how I missed this one. Tortle playing as a priest wouldn't be
able to wear its own quest artifact (Mitre of Holiness) because it's
made of a hard/rigid material. So I made an exception only for that
object, and only if the tortle is a priest.

While working on this, found a few other bugs. Example - say you're a
tortle, and you polymorph into a humanoid monster that could wear boots.
You still wouldn't be able to put those boots on. This affected centaurs
also. And there were a few other cases as well. So I fixed those.

Which set the stage for another bug. Same example - you're a centaur,
you polymorph into a humanoid monster that can wear boots. You put on
some boots, and eventually you revert to back to your centaur form. Your
boots are still being worn. So I added a call to break_armor() in
rehumanize(), which should have been done a long time ago with the
inclusion of giant race.

So, killed three birds with one stone 😀