Allow helm of speed to be made of more materials.
When I created the centaur player race, I created the helm of speed
since they can't wear boots (speed boots specifically). Now we have the
tortle race, that can wear hats but nothing made of a rigid material,
and the helm of speed has its base material as steel, and could only
form as other hard materials. This limits usable prize items for a
tortle attempting to complete sokoban.

Solution - make a special case for the helm of speed, allow it to spawn
as leather, cloth, or dragonhide, as well as all the other materials it
can be made from. Its base material is still steel, and has a higher
chance of being that than any other material, but there's a significant
chance the sokoban prize helm will be of a material a tortle can use. Or
they could always wish one up if they so chose.