Give tortles a slightly higher weight carry cap.
Tortles can reach 19 strength naturally, are larger than humans, and lug
around a huge shell on their back 24/7, so have their carry cap reflect
their natural strength (additional 200 aum). Not as great as a giant or
centaur, but still respectable.

Also noticed that I had the weight wrong for a tortle in monst.c, it was
actually less than a human, so fixed that as well.